Moved booru command to it's own file
Characters are unescaped in tags reply.
Shorthand command "!b" added as well as a "rating:safe" option in "!sb"
